Bug Fixes for #0002674 & #0002659
Add button named 'Add as Friend' below the user profile page if viewed
user is not yet a friend of the viewing user.

Change also the label of the add button of search friend results.

During registration, there should be a validation that will check for
the values [data entry] for email, name, password, above 18.

The application should only continue with the registration if the said
fields are entered correctly.